+++ src/main/java/de/rwth/idsg/steve/CentralSystemService15_Impl.java
@@ -35,6 +35,7 @@ import ocpp.cs._2012._06.StopTransactionResponse;
 import ocpp.cs._2012._06.TransactionData;
 
 import org.apache.cxf.ws.addressing.AddressingProperties;
+import org.joda.time.DateTime;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -70,6 +71,8 @@ public class CentralSystemService15_Impl implements CentralSystemService {
 		AddressingProperties addressProp = (AddressingProperties) messageContext.get(org.apache.cxf.ws.addressing.JAXWSAConstants.SERVER_ADDRESSING_PROPERTIES_INBOUND);
 		String endpoint_address = addressProp.getFrom().getAddress().getValue();
 		
+		DateTime now = new DateTime();
+		
 		boolean isRegistered = ServiceDBAccess.updateChargebox(endpoint_address,
 				"1.5",
 				parameters.getChargePointVendor(),
@@ -81,14 +84,15 @@ public class CentralSystemService15_Impl implements CentralSystemService {
 				parameters.getImsi(),
 				parameters.getMeterType(),
 				parameters.getMeterSerialNumber(),
-				chargeBoxIdentity);
+				chargeBoxIdentity, 
+				new Timestamp(now.getMillis()));
 		
 		BootNotificationResponse _return = new BootNotificationResponse();
 		RegistrationStatus _returnStatus = null;
 		
 		if (isRegistered) {
 			_returnStatus = RegistrationStatus.ACCEPTED;
-			_return.setCurrentTime(DateTimeUtils.getCurrentDateTimeXML());
+			_return.setCurrentTime(DateTimeUtils.convertToXMLGregCal(now));
 			_return.setHeartbeatInterval(Integer.valueOf(Constants.HEARTBEAT_INTERVAL));
 		} else {
 			_returnStatus = RegistrationStatus.REJECTED;
@@ -209,9 +213,12 @@ public class CentralSystemService15_Impl implements CentralSystemService {
 
 	public HeartbeatResponse heartbeat(HeartbeatRequest parameters,java.lang.String chargeBoxIdentity) { 
 		LOG.info("Executing heartbeat for {}", chargeBoxIdentity);
+		
+		DateTime now = new DateTime();
+		ServiceDBAccess.updateChargeboxHeartbeat(chargeBoxIdentity, new Timestamp(now.getMillis()));
 
 		HeartbeatResponse _return = new HeartbeatResponse();
-		_return.setCurrentTime(DateTimeUtils.getCurrentDateTimeXML());
+		_return.setCurrentTime(DateTimeUtils.convertToXMLGregCal(now));
 		return _return;
 	}