+++ src/main/webapp/WEB-INF/jsp/00-js-snippets/getCPDetails.js
@@ -1,0 +1,15 @@
+$(document).ready(function() {
+    $("#gdb").click(function() {
+		var ddiv = $("#details-div");
+		ddiv.html("<br>Loading...");		
+    	$.getJSON("${contextPath}/manager/ajax/getCPDetails?chargeBoxId=" + $("#cbi").val(), function(data) {
+    		ddiv.html("<table id='details' class='cpd'><thead><tr><th>Charge Point Details</th><th></th></tr></thead></table>");
+    		var table = $("#details");
+            $.each(data, function(key, val) {
+	            $("<tr>").appendTo(table)
+		            .append($("<td>").text(key))
+		            .append($("<td>").text(val));
+            });
+    	});
+    });
+});