+++ src/main/webapp/WEB-INF/jsp/00-js-snippets/getConnectorIdsZeroAllowed.js
@@ -1,0 +1,12 @@
+$("#cp_items").change(function() {
+	var cp = $(this).val().toString().split(';')[0];
+	$.getJSON("${contextPath}/manager/ajax/getConnectorIds?chargeBoxId=" + cp, function(data) {	
+		var options = "<option value='0'>Not for a specific connector</option>";
+		$.each(data, function() {
+			options += "<option value='" + this + "'>" + this + "</option>";
+		});
+		var select = $("#connectorId");
+		select.prop("disabled", false);
+		select.html(options);
+	});
+});